requirements.txt
automium==0.2.6
automium-web==0.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.4
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
legacy==0.1.6
MarkupSafe==1.1.1
pytz==2019.2
requests==2.22.0
reverse==0.1.0
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.3
Werkzeug==0.15.5
whitenoise==4.1.4
django-paypal==1.0.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
